requirements.txt
appdirs==1.4.4
asgiref==3.3.4
astroid==2.5.3
colorama==0.4.4
distlib==0.3.1
Django==3.0.5
django-cors-headers==3.7.0
djangorestframework==3.12.4
djongo==1.3.4
filelock==3.0.12
isort==5.8.0
lazy-object-proxy==1.6.0
mccabe==0.6.1
pylint==2.7.4
pymongo==3.11.3
pytz==2021.1
six==1.15.0
sqlparse==0.2.4
toml==0.10.2
virtualenv==20.4.3
wrapt==1.12.1
